+++ src/app/jobsite2/layout.js
@@ -1,5 +1,5 @@
-import Top2 from '@/components/jobsite2/Top2';
-import Footer from '@/components/Footer';
+import Top from '@/components/jobsite/Top';
+import Footer from '@/components/jobsite/Footer';
 
 export const metadata = {
   title: 'New Comers Job Site',
@@ -9,7 +9,7 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body>
-        <Top2 />
+        <Top />
         {children}
         <Footer />
       </body>
